ai-service/requirements.txt
# Dependências do microserviço Flask
flask==2.3.3
flask-cors==4.0.0
cohere==4.39
python-dotenv==1.0.0
gunicorn==21.2.0
